requirements.txt
tensorflow
numpy
Pillow
matplotlib
streamlit
